+++ cli/azd/internal/vsrpc/environment_service_load.go
@@ -60,7 +60,10 @@ func (s *environmentService) loadEnvironmentAsyncWithSession(
 		dotnetCli     dotnet.DotNetCli       `container:"type"`
 	}
 
-	container := newContainer(session)
+	container, err := session.newContainer()
+	if err != nil {
+		return nil, err
+	}
 	if err := container.Fill(&c); err != nil {
 		return nil, err
 	}
